BM: Cleanups, move disable profiling code to -extras.cpp
