+++ core/hw/sh4/dyna/blockmanager.h
@@ -110,3 +110,5 @@ void bm_Init();
 void bm_Term();
 
 void bm_vmem_pagefill(void** ptr,u32 PAGE_SZ);
+
+void bm_sh4_jitsym(FILE* out);
